refactor: 기능 분리 (InputView 생성)
