+++ src/Controllers/CarController.js
@@ -1,4 +1,28 @@
+import { InputView } from '../Views/InputView.js';
 class CarController {
+  run() {
+    const carNamesInput = this.getCarName();
+    const carNames = this.validateCarName();
+
+    // const tryCountInput = await readLineAsync('시도할 횟수는 몇 회인가요?\n');
+    // const tryCount = stringToNumber(tryCountInput);
+    // checkTryCountRange(tryCount);
+    // checkIsInteger(tryCount)
+  }
+
+  getCarName() {
+    return InputView.inputCarName();
+  }
+
+  validateCarName() {
+    checkIsEmpty(carNamesInput);
+    const carNames = splitString(carNamesInput);
+    checkCarNameLength(carNames);
+    checkCarCount(carNames);
+    checkCarNameDuplicate(carNames);
+    return carNames;
+  }
+
   getRandomNumber() {
     return Math.floor(Math.random() * 10);
   }
