+++ src/Views/InputView.js
@@ -1,0 +1,7 @@
+class InputView {
+  static async inputCarName() {
+    return await readLineAsync('경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).\n');
+  }
+}
+
+export default InputView;
